+++ Character.h
@@ -6,19 +6,19 @@ class Character
     private:   
         std::string Name;
         int Level;
-        double MaxHP, HP, ATK, DEF;
+        int MaxHP, HP, ATK, DEF;
     public:
-        Character(std::string, int, double, double, double);
+        Character(std::string, int, int, int, int);
         std::string getName();
         void setLevel(int);
         int getLevel();
-        void setMaxHP(double);
-        double getMaxHP();
-        void setHP(double);
-        double getHP();
-        void setATK(double);
-        double getATK();
-        void setDEF(double);
-        double getDEF();
+        void setMaxHP(int);
+        int getMaxHP();
+        void setHP(int);
+        int getHP();
+        void setATK(int);
+        int getATK();
+        void setDEF(int);
+        int getDEF();
 };
 #endif
