+++ Player.h
@@ -8,14 +8,14 @@ class Enemy;
 class Player: public Character
 {
     private:
-        double XP;
+        int XP;
         // Weapon *equippedWeapon;
         // Item *Player_Item;
         long Money;
         int SP;
         std::string Player_Class;
     public:
-        Player(std::string, double, double, double, std::string);
+        Player(std::string, int, int, int, std::string);
         void Show_Status();
         void Attack(Enemy&);
         virtual void Skill(Enemy&);
@@ -23,7 +23,7 @@ class Player: public Character
         //void Using_Item(); //chắc truy xuất tới hàm dùng item của cái class item
         void Show_Description();
         void Show_Skill_Description();
-        void Gain_XP(double);
+        void Gain_XP(int);
         //void equipWeapon(Weapon&);
         //void useItem(int itemID);
         friend class Enemy;
